+++ cpp/subprojects/boosting/include/boosting/binning/label_binning_no.hpp
@@ -36,12 +36,18 @@ namespace boosting {
             std::unique_ptr<ILabelWiseRuleEvaluationFactory> createLabelWiseFixedPartialRuleEvaluationFactory(
                 float32 labelRatio, uint32 minLabels, uint32 maxLabels) const override;
 
+            std::unique_ptr<ILabelWiseRuleEvaluationFactory> createLabelWiseDynamicPartialRuleEvaluationFactory(
+                float32 threshold, float32 exponent) const override;
+
             std::unique_ptr<IExampleWiseRuleEvaluationFactory> createExampleWiseCompleteRuleEvaluationFactory(
                 const Blas& blas, const Lapack& lapack) const override;
 
             std::unique_ptr<IExampleWiseRuleEvaluationFactory> createExampleWiseFixedPartialRuleEvaluationFactory(
                 float32 labelRatio, uint32 minLabels, uint32 maxLabels, const Blas& blas,
                 const Lapack& lapack) const override;
+
+            std::unique_ptr<IExampleWiseRuleEvaluationFactory> createExampleWiseDynamicPartialRuleEvaluationFactory(
+                float32 threshold, float32 exponent, const Blas& blas, const Lapack& lapack) const override;
 
     };