+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 A major update to the BOOMER algorithm that introduces the following changes:
 
-* The parameter `--head-type` does now allow to learn partial rules that predict for a predefined number of labels (`partial-fixed`).
+* The parameter `--head-type` does now allow to learn partial rules that predict for a predefined number of labels (`partial-fixed`) or a subset of the available labels that is determined dynamically (`partial-dynamic`).
 * A new parameter `--probability-predictor` has been added. It allows to specify whether probability estimates should be obtained for each label independently (`label-wise`) or via marginization over the label vectors that are encountered in the training data (`marginalized`).
 
 ### Version 0.8.2 (Apr. 11th, 2022)